First run on Furniture!
Added everything from the Furniture add-on.  The tea table and stool are great!  The chair is what it is.  Could probably improve it some with directional placement for 'down'. 
The bench and mirror are WIP.  I can make them work.
The bookshelves are beautiful, but MASSSIVE!

All the textures are wonderful!