This is a development build.  Have fun.
This build has **lots** of console output and several things don't work exactly right.
The item_orange is food and canEatAlways() is set to true.  The block_orange can be placed, but it always drops the food item_orange.  I'm still working on that.
The *right-click to pick up things* was working great.  Except it added to an existing stack in your inventory, which may be in your hand or not.  Currently it places the item in your hand, but doesn't check to see if it's empty.  There's also a duplication error if you have multiple empty slots in your hotbar.  I'm working on that, too.
This is not a public release, I just thought you might want to play with the directional placement and see the fixed texture.

Have fun.